[#10] Cached theme meta. This resolves the error ("can't call method 'registry'") by caching information and using additional fallback options.
This also adds a feature to the Production Mode: cached data is returned, resulting in a clearer view of the currently installed theme. The theme version number of the installed theme, for example, is saved. Previously, installing a newer version of a theme would result in confusion on the Theme Dashboard where the newer theme version number is displayed, while the older theme is actually what is installed.

Designer Mode will continue to load the theme data dynamically.